// dash/src/types/state.type.ts
export type DriverList = {
	[key: string]: Driver;
};

export type Driver = {
	racingNumber: string;
	broadcastName: string;
	fullName: string;
	tla: string;
	line: number;
	teamName: string;
	teamColour: string;
	firstName: string;
	lastName: string;
	reference: string;
	headshotUrl: string;
	countryCode: string;
};
// ...
export type TimingDataDriver = {
	stats?: { timeDiffToFastest: string; timeDifftoPositionAhead: string }[];
	timeDiffToFastest?: string;
	timeDiffToPositionAhead?: string;
	gapToLeader: string;
	intervalToPositionAhead?: {
		value: string;
		catching: boolean;
	};
	line: number;
	position: string;
	showPosition: boolean;
	racingNumber: string;
	retired: boolean;
	inPit: boolean;
	pitOut: boolean;
	stopped: boolean;
	status: number;
	sectors: Sector[];
	speeds: Speeds;
	bestLapTime: PersonalBestLapTime;
	lastLapTime: I1;
	numberOfLaps: number; // TODO check
	knockedOut?: boolean;
	cutoff?: boolean;
};

export type Sector = {
	stopped: boolean;
	value: string;
	previousValue?: string;
	status: number;
	overallFastest: boolean;
	personalFastest: boolean;
	segments: {
		status: number;
	}[];
};

export type Speeds = {
	i1: I1;
	i2: I1;
	fl: I1;
	st: I1;
};

export type I1 = {
	value: string;
	status: number;
	overallFastest: boolean;
	personalFastest: boolean;
};

export type TimingStatsDriver = {
	line: number;
	racingNumber: string;
	personalBestLapTime: PersonalBestLapTime;
	bestSectors: PersonalBestLapTime[];
	bestSpeeds: {
		i1: PersonalBestLapTime;
		i2: PersonalBestLapTime;
		fl: PersonalBestLapTime;
		st: PersonalBestLapTime;
	};
};

export type PersonalBestLapTime = {
	value: string;
	position: number;
};